Back to Int to see

+++ src/ArchimedeanCopula.jl
@@ -6,7 +6,7 @@ Fields:
 
 Constructor:
 
-    ArchimedeanCopula(d::Int64,G::Generator)
+    ArchimedeanCopula(d::Int,G::Generator)
 
 For some Archimedean [`Generator`](@ref) `G::Generator` and some dimenson `d`, this class models the archimedean copula which has this generator. The constructor checks for validity by ensuring that `max_monotony(G) ≥ d`. The ``d``-variate archimedean copula with generator ``\\phi`` writes:
 
@@ -43,13 +43,13 @@ References:
 """
 struct ArchimedeanCopula{d,TG} <: Copula{d}
     G::TG
-    function ArchimedeanCopula(d::Int64,G::Generator)
+    function ArchimedeanCopula(d::Int,G::Generator)
         @assert d <= max_monotony(G) "The generator $G you provided is not $d-monotonous since it has max monotonicity $(max_monotony(G)), and thus this copula does not exists."
         return new{d,typeof(G)}(G)
     end
-    ArchimedeanCopula(d::Int64, ::IndependentGenerator) = IndependentCopula(d)
-    ArchimedeanCopula(d::Int64, ::MGenerator) = MCopula(d)
-    ArchimedeanCopula(d::Int64, ::WGenerator) = WCopula(d)
+    ArchimedeanCopula(d::Int, ::IndependentGenerator) = IndependentCopula(d)
+    ArchimedeanCopula(d::Int, ::MGenerator) = MCopula(d)
+    ArchimedeanCopula(d::Int, ::WGenerator) = WCopula(d)
     ArchimedeanCopula{d,TG}(θ) where {d, TG} = ArchimedeanCopula(d, TG(θ))
 end
 Distributions.params(C::ArchimedeanCopula) = Distributions.params(C.G) # by default the parameter is the generator's parameters. 
@@ -159,7 +159,7 @@ function inverse_rosenblatt(C::ArchimedeanCopula{d,TG}, u::AbstractMatrix{<:Real
 end
 
 # Conditioning colocated
-function DistortionFromCop(C::ArchimedeanCopula, js::NTuple{p,Int64}, uⱼₛ::NTuple{p,Float64}, i::Int64) where {p}
+function DistortionFromCop(C::ArchimedeanCopula, js::NTuple{p,Int}, uⱼₛ::NTuple{p,Float64}, i::Int) where {p}
     @assert length(js) == length(uⱼₛ)
     sJ = zero(eltype(uⱼₛ))
     @inbounds for u in uⱼₛ
@@ -169,9 +169,9 @@ function DistortionFromCop(C::ArchimedeanCopula, js::NTuple{p,Int64}, uⱼₛ::N
 end
 
 # Conditional copula specialization: remains Archimedean with a tilted generator
-function ConditionalCopula(C::ArchimedeanCopula{D}, ::NTuple{p,Int64}, uⱼₛ::NTuple{p,Float64}) where {D, p}
+function ConditionalCopula(C::ArchimedeanCopula{D}, ::NTuple{p,Int}, uⱼₛ::NTuple{p,Float64}) where {D, p}
     return ArchimedeanCopula(D - p, TiltedGenerator(C.G, Val{p}(), sum(ϕ⁻¹.(C.G, uⱼₛ))))
 end
 
 # Subsetting colocated
-SubsetCopula(C::ArchimedeanCopula{d,TG}, dims::NTuple{p, Int64}) where {d,TG,p} = ArchimedeanCopula(length(dims), C.G)
+SubsetCopula(C::ArchimedeanCopula{d,TG}, dims::NTuple{p, Int}) where {d,TG,p} = ArchimedeanCopula(length(dims), C.G)
